pyproject.toml
[project]
name = "env-loader"
version = "0.1.0"
description = ""
authors = [
    { name = "Seika139", email = "33862136+Seika139@users.noreply.github.com" },
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
license = { file = "LICENSE" }
dependencies = ["python-dotenv (>=1.1.0,<2.0.0)"]

[tool.poetry]
packages = [{ include = "env_loader", from = "src" }]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
mypy = "^1.15.0"
pytest = "^8.3.5"
isort = "^6.0.1"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.flake8]
max-line-length = 88
# Ignore Errors which can conflict with black
extend-ignore = ["E203", "E501", "F401", "W503"]
# Check explicitly for Errors (E), Warnings (W), and flake8 Errors (F)
select = ["E", "W", "F"]

[tool.mypy]
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
strict_optional = true
namespace_packages = true
explicit_package_bases = true
mypy_path = ["src"]
